Feature/controller layer

+++ src/main/java/denis/orderservice/service/impl/OrderServiceImpl.java
@@ -1,6 +1,7 @@
 package denis.orderservice.service.impl;
 
 import denis.orderservice.client.UserServiceClient;
+import denis.orderservice.dto.request.OrderItemRequestDto;
 import denis.orderservice.dto.request.OrderRequestDto;
 import denis.orderservice.dto.request.OrderUpdateRequestDto;
 import denis.orderservice.dto.response.OrderResponseDto;
@@ -9,9 +10,12 @@ import denis.orderservice.entity.Item;
 import denis.orderservice.entity.Order;
 import denis.orderservice.entity.OrderItem;
 import denis.orderservice.entity.OrderStatus;
+import denis.orderservice.exception.ItemNotFoundException;
+import denis.orderservice.exception.OrderNotFoundException;
 import denis.orderservice.mapper.OrderItemMapper;
 import denis.orderservice.mapper.OrderMapper;
 import denis.orderservice.repository.ItemRepository;
+import denis.orderservice.repository.OrderItemRepository;
 import denis.orderservice.repository.OrderRepository;
 import denis.orderservice.repository.specification.OrderSpecifications;
 import denis.orderservice.service.OrderService;
@@ -20,9 +24,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -30,17 +34,19 @@ import java.util.UUID;
 public class OrderServiceImpl implements OrderService {
 
     private final OrderRepository orderRepository;
+    private final OrderItemRepository orderItemRepository;
     private final ItemRepository itemRepository;
     private final OrderMapper orderMapper;
     private final OrderItemMapper orderItemMapper;
     private final UserServiceClient userClient;
 
-    public OrderServiceImpl(OrderRepository orderRepository,
+    public OrderServiceImpl(OrderRepository orderRepository, OrderItemRepository orderItemRepository,
                             ItemRepository itemRepository,
                             OrderMapper orderMapper,
                             OrderItemMapper orderItemMapper,
                             UserServiceClient userClient) {
         this.orderRepository = orderRepository;
+        this.orderItemRepository = orderItemRepository;
         this.itemRepository = itemRepository;
         this.orderMapper = orderMapper;
         this.orderItemMapper = orderItemMapper;
@@ -49,14 +55,24 @@ public class OrderServiceImpl implements OrderService {
 
     @Transactional
     @Override
-    public OrderResponseDto  create(OrderRequestDto dto) {
+    public OrderResponseDto create(OrderRequestDto dto) {
         Order order = orderMapper.toEntity(dto);
+        order.setStatus(OrderStatus.CREATED);
+        order.setItems(new ArrayList<>());
+        return saveCreateOrderItem(order,dto.items());
+    }
+
+    public OrderResponseDto saveCreateOrderItem(Order order,List<OrderItemRequestDto> itemsDto){
         BigDecimal total = BigDecimal.ZERO;
-        for (OrderItem oi : order.getItems()) {
-            Item item = itemRepository.findById(oi.getItem().getId())
-                    .orElseThrow(() -> new IllegalArgumentException("Item not found: " + oi.getItem().getId()));//написать свою exc
-            oi.setItem(item);
-            total = total.add(item.getPrice().multiply(BigDecimal.valueOf(oi.getQuantity())));
+        for (OrderItemRequestDto oi : itemsDto) {
+            Item item = itemRepository.findById(oi.itemId())
+                    .orElseThrow(() -> new ItemNotFoundException("Item not found: " + oi.itemId()));
+            OrderItem orderItem = new OrderItem();
+            orderItem.setQuantity(oi.quantity());
+            orderItem.setItem(item);
+            orderItem.setOrder(order);
+            order.getItems().add(orderItem);
+            total = total.add(item.getPrice().multiply(BigDecimal.valueOf(oi.quantity())));
         }
         order.setTotalPrice(total);
         Order saved = orderRepository.save(order);
@@ -68,7 +84,7 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public OrderResponseDto  getById(UUID id) {
         Order order = orderRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Order not found"));//написать свою exc
+                .orElseThrow(() -> new OrderNotFoundException("Order not found"));
         UserInfoDto user = userClient.getUserById(order.getUserId());
         return orderMapper.toDto(order, user);
     }
@@ -103,28 +119,18 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public OrderResponseDto update(UUID id, OrderUpdateRequestDto dto) {
         Order order = orderRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Order not found"));//написать свою exc
+                .orElseThrow(() -> new OrderNotFoundException("Order not found"));
         order.setStatus(dto.status());
         order.getItems().clear();
-        List<OrderItem> updatedItems = dto.items().stream().map(orderItemMapper::toEntity).toList();
-
-        BigDecimal total = BigDecimal.ZERO;
-        for (OrderItem oi : updatedItems) {
-            Item item = itemRepository.findById(oi.getItem().getId())
-                    .orElseThrow(() -> new IllegalArgumentException("Item not found: " + oi.getItem().getId()));//написать свою exc
-            oi.setItem(item);
-            total = total.add(item.getPrice().multiply(BigDecimal.valueOf(oi.getQuantity())));
-        }
-        order.setTotalPrice(total);
-        order.getItems().addAll(updatedItems);
-        Order saved = orderRepository.save(order);
-        UserInfoDto user = userClient.getUserById(order.getUserId());
-        return orderMapper.toDto(saved, user);
+        return saveCreateOrderItem(order, dto.items());
     }
 
     @Transactional
     @Override
     public void delete(UUID id) {
+        if (!orderRepository.existsById(id)) {
+            throw new OrderNotFoundException("Order not found with id: " + id);
+        }
         orderRepository.deleteById(id);
     }
 }
